@@ -4,13 +4,13 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
+  charts_flutter: ^0.12.0
   flutter:
     sdk: flutter
-  charts_flutter: ^0.12.0
-  material_design_icons_flutter: ^4.0.5045
+  material_design_icons_flutter: ^7.0.7296
   
   sliding_sheet:
     path: ../
