@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:charts_flutter/flutter.dart' as charts;
 import 'package:example/util/util.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
@@ -166,8 +165,8 @@ class _ExampleState extends State<Example> {
       Icon icon,
       Text text,
       VoidCallback onTap, {
-      BorderSide border,
-      Color color,
+      BorderSide? border,
+      Color? color,
     }) {
       final child = Row(
         mainAxisSize: MainAxisSize.min,
@@ -217,7 +216,7 @@ class _ExampleState extends State<Example> {
             ),
             () async {
               // Inherit from context...
-              await SheetController.of(context).hide();
+              await SheetController.of(context)?.hide();
               Future.delayed(const Duration(milliseconds: 1500), () {
                 // or use the controller
                 controller.show();
@@ -526,7 +525,7 @@ class _ExampleState extends State<Example> {
                 children: <Widget>[
                   Text(
                     'Confirm purchase',
-                    style: textTheme.headline4.copyWith(
+                    style: textTheme.headlineMedium?.copyWith(
                       color: Colors.white,
                       fontWeight: FontWeight.bold,
                     ),
@@ -537,7 +536,7 @@ class _ExampleState extends State<Example> {
                       Expanded(
                         child: Text(
                           'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent sagittis tellus lacus, et pulvinar orci eleifend in.',
-                          style: textTheme.subtitle1.copyWith(
+                          style: textTheme.titleMedium?.copyWith(
                             color: Colors.white,
                             fontWeight: FontWeight.bold,
                           ),
@@ -566,7 +565,7 @@ class _ExampleState extends State<Example> {
                     onPressed: () => Navigator.pop(context),
                     child: Text(
                       'Cancel',
-                      style: textTheme.subtitle1.copyWith(
+                      style: textTheme.titleMedium?.copyWith(
                         color: Colors.white,
                         fontWeight: FontWeight.bold,
                       ),
@@ -577,14 +576,14 @@ class _ExampleState extends State<Example> {
                     onPressed: () {
                       if (!isDismissable) {
                         isDismissable = true;
-                        SheetController.of(context).rebuild();
+                        SheetController.of(context)?.rebuild();
                       } else {
                         Navigator.pop(context);
                       }
                     },
                     child: Text(
                       'Approve',
-                      style: textTheme.subtitle1.copyWith(
+                      style: textTheme.titleMedium?.copyWith(
                         color: Colors.white,
                         fontWeight: FontWeight.bold,
                       ),
